+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.7.0
+* Updated boilerplate extension code endpoint
+* Updated to use `@shopgate/cloud-sdk-webpack@1.17.1`
+* Added `extension upload` and `theme upload` commands for uploading extensions and themes respectively
+* Updated error handling for various occasions e.g. wrong organisation
+* Added extension-config.json validation on backend start
+
 ## v1.6.0
 * Promisified version of `tracedRequest` method
 * Upgraded vulnerable dependencies